Move defaultRoles to PropTypes
