+++ src/Caption.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import LocaleUtils from './LocaleUtils';
 
 import { ENTER } from './keys';
-import { RoleTypesShape, defaultRoles } from './DayPicker';
+import { RoleTypesShape, defaultRoles } from './PropTypes';
 
 export default class Caption extends Component {
   static propTypes = {
@@ -21,7 +21,6 @@ export default class Caption extends Component {
 
   static defaultProps = {
     localeUtils: LocaleUtils,
-    roles: defaultRoles,
   };
 
   constructor(props) {
@@ -52,7 +51,7 @@ export default class Caption extends Component {
       locale,
       localeUtils,
       onClick,
-      roles,
+      roles = defaultRoles,
     } = this.props;
     return (
       <div className={classNames.caption} role={roles.caption}>
